requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
djangorestframework==3.15.1
filelock==3.13.4
fsspec==2024.3.1
idna==3.7
packaging==24.0
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1